FIX prediction depth: separate query points in KNN
FIX add 2 new arguments to prediction_depth for the query points
(intermediate activations and labels) when fitting the KNN classifier.
This allows the query points to differ from the points to be classified.
In particular, Baldock et al. (2021) use the training dataset
as the query points, and the test dataset as the points to predict.
The issue with reusing test points as queries is that
a point's nearest neighbour is always itself.
This problem is most notable when k=1.