CHANGE make add() in-place for online mean/var
CHANGE for Accumulator objects,
previously add() would return a new object, guaranteeing immutability
However, this is counter to intuition on how users expect add() to work.
Now add() will modify the object in place but also return itself
The only change is that get() no longer returns an immutable tensor,
since it returns a reference to the object
Future calls to add() will change the tensor given by get.
Since it is possible to use Accumulator objects for gradient computations,
it is up to the end user to clone the output of get() when needed.